[manuf] move CP provisioning harness code to a lib
This moves the CP provisioning Rust harness code to a lib so it can be reused
by a functest that tests the CP provisioning binary itself, and a Rust
harness that simply drives the CP provisioning SRAM binary to
provision ES devices.
