+++ sw/host/tests/manuf/cp_provision_init/BUILD
@@ -3,43 +3,17 @@
 # SPDX-License-Identifier: Apache-2.0
 
 load("@rules_rust//rust:defs.bzl", "rust_binary")
-load("//rules:ujson.bzl", "ujson_rust")
-load("//rules:lc.bzl", "lc_raw_unlock_token")
 
 package(default_visibility = ["//visibility:public"])
-
-lc_raw_unlock_token(
-    name = "lc_raw_unlock_token",
-)
-
-ujson_rust(
-    name = "provisioning_command",
-    srcs = ["//sw/device/lib/testing/json:provisioning_command"],
-)
-
-ujson_rust(
-    name = "provisioning_data",
-    srcs = ["//sw/device/lib/testing/json:provisioning_data"],
-)
 
 rust_binary(
     name = "cp_provision_init",
     srcs = [
         "src/main.rs",
-        "src/provisioning_command.rs",
-        "src/provisioning_data.rs",
-        ":lc_raw_unlock_token",
     ],
-    compile_data = [
-        ":provisioning_command",
-        ":provisioning_data",
-    ],
-    rustc_env = {
-        "provisioning_command": "$(location :provisioning_command)",
-        "provisioning_data": "$(location :provisioning_data)",
-    },
     deps = [
         "//sw/host/opentitanlib",
+        "//sw/host/provisioning/cp_lib",
         "@crate_index//:anyhow",
         "@crate_index//:arrayvec",
         "@crate_index//:clap",
